+++ src/App.tsx/App.tsx
@@ -10,7 +10,6 @@ import Layout from '../layout/Layout';
 import { offersSelectors } from '@/storage/slices/offers';
 import { fetchOffers } from '@/thunk/offers';
 import { useEffect } from 'react';
-import { authorisationSelectors } from '@/storage/slices/authorization';
 import { useAppDispatch, useAppSelector } from '@/hooks';
 import { checkAuthorisation } from '@/thunk/authorisation';
 
@@ -21,8 +20,6 @@ function App() {
   const offers = useAppSelector(offersSelectors.selectOffers);
 
   const activeCity = useAppSelector(offersSelectors.selectActiveCity);
-
-  const authState = useAppSelector(authorisationSelectors.selectAuthorisationStatus);
 
 
   useEffect(() => {
@@ -37,7 +34,7 @@ function App() {
         <Route index path={`${AppRoutes.Main}`} element={<MainPage activeCity={activeCity} offers={offers}/>} />
         <Route index path={`${AppRoutes.Main}:city`} element={<MainPage activeCity={activeCity} offers={offers}/>} />
         <Route path={AppRoutes.Favourites} element={
-          <PrivateRoute authorisationStatus={authState}><FavouritesPage /></PrivateRoute>
+          <PrivateRoute><FavouritesPage /></PrivateRoute>
         }
         />
         <Route path={AppRoutes.Offers} element={<OfferPage offers={offers} />} />
