+++ src/components/private-route/private-route.tsx
@@ -1,12 +1,15 @@
 import { Navigate } from 'react-router-dom';
 import { AppRoutes, AuthorisationStatus } from '../../libs/const';
+import { authorisationSelectors } from '@/storage/slices/authorization';
+import { useAppSelector } from '@/hooks';
 
 type PrivateRouteProps = {
-  authorisationStatus: AuthorisationStatus;
   children: JSX.Element;
 }
 
-function PrivateRoute({authorisationStatus, children}: PrivateRouteProps) {
+function PrivateRoute({ children}: PrivateRouteProps) {
+
+  const authorisationStatus = useAppSelector(authorisationSelectors.selectAuthorisationStatus);
   return (
     authorisationStatus === AuthorisationStatus.Auth ? children :
       <Navigate to={AppRoutes.Login}/>
